+++ iti-react/src/Inputs/Date/DateInputNoPicker.tsx
@@ -133,22 +133,21 @@ export const DateInputNoPicker = React.memo<DateInputNoPickerProps>(
             validationKey: otherProps.validationKey,
             asyncValidator: otherProps.asyncValidator,
             onAsyncError: otherProps.onAsyncError,
-            onAsyncValidationInProgressChange:
-                otherProps.onAsyncValidationInProgressChange,
             formLevelValidatorOutput: otherProps.formLevelValidatorOutput,
         })
 
-        const classes = ['form-control', getValidationClass(valid, showValidation)]
+        const classes = [
+            'form-control',
+            getValidationClass(!validatorOutput, showValidation),
+        ]
         if (otherProps.className) classes.push(otherProps.className)
 
         const className = classes.join(' ')
 
         return (
             <ValidationFeedback
-                valid={valid}
+                validatorOutput={validatorOutput}
                 showValidation={showValidation}
-                invalidFeedback={invalidFeedback}
-                asyncValidationInProgress={asyncValidationInProgress}
             >
                 <input
                     id={id}