+++ iti-react/src/Inputs/RadioInput/RadioInput.tsx
@@ -49,7 +49,6 @@ export const RadioInput = React.memo<RadioInputProps>(
             validationKey: props.validationKey,
             asyncValidator: props.asyncValidator,
             onAsyncError: props.onAsyncError,
-            onAsyncValidationInProgressChange: props.onAsyncValidationInProgressChange,
             formLevelValidatorOutput: props.formLevelValidatorOutput,
         })
 
@@ -63,10 +62,8 @@ export const RadioInput = React.memo<RadioInputProps>(
 
         return (
             <ValidationFeedback
+                validatorOutput={validatorOutput}
                 showValidation={showValidation}
-                valid={valid}
-                invalidFeedback={invalidFeedback}
-                asyncValidationInProgress={asyncValidationInProgress}
             >
                 <div className={containerClasses.join(' ')}>
                     {options.map((o) => (